EOC : final tweaks
Ensured image is now visible, was due to the constraints for width being missing. I have changed the constraints to limit the image from an arbitrary number from each view each to ensure the image is stretched across the view as required. I have used Aspect Fit to ensure the image isn't skewed. Also added hidesBarOnTap in the detailViewController.